(Creating layout page #9) In "mounted" lifecycle hook: created and called 'swapLogosWhileRotating' function.
Function description: I essentialy used some JS here to "swap"
the fragments of the logo to be hidden (namely: the 'home' and 'fair' svgs)
when they're rotating so that it looks like they're transformed
into the 'knot' and 'bond' svg during the rotation.
I provided the "IDEA" tag comment to describe it in the code
should you wish to read more.