(Creating layout page #14) Added padding to the bottom of the main background window.
Please note whenever I use the "!important" in the styles
it is to override the default styling of the Vuetify's built-in
components. Vuetify is great with it's support of the vue
directives on their elements, but it's also quite moody to
work with when we want to operate 'deeper' than the
directives level and actually change its styling.
Thus, as much as I avoid the !importants, it's rather a
necessity with Vuetify and how its internal styling is made.