(FINAL: Creating layout page #16) Added background image to the website's main body.
Note that due to time constraints I didn't optimise the load
of the image and just "stuffed the image and let it be"
into the website. Normally I would convert the image to .webp
format and configure webpack to process it efficiently.
Also, probably even more importantly, I would
tinker with the ':srcset' attribute if it was an html tag
or add other media queries for smaller-size images
on smaller devices.
FINALLY: this concludes the work on the default layout page.