(Creating Vuex store #5) Created "getFixedMembershipInfo" vuex action.
Function/action description:
1. This func. sets the loading state to true,
does a GET request to the AWS URL endpoint
and obtains the response object:  like the following: {
    "fixed_membership_fee": true,
    "fixed_membership_fee_amount": 6781
}
2. Upon receival, it sets the loading state back to false and
also calls the setInfoAboutFixedMembershipFee (described in the previous commit).
3. Notice the way I call the axios module is non-standard. That's because I used
@Nuxtjs/axios version of axios - it's supposedly better as it offers some deeper
integration with nuxtjs node-based server that's running in background
(integration with the nuxt's loading bar while fetching seems nifty !) but frankly
I didn't observe any of those extra features by using it out of the box :>
Still, it's a 'potential' win if properly configured.