(Creating Vuex store #7) Created "setFlatbondDetails" mutation.
Mutation description:
1. This mut. sets the FlatbondDetails object
with all the essential bits of info we gather
from the user form on the first page.
2. Additionally, we check what's the value of
"weeklyOrMonthlyValue" variable we will pass from the
index.vue (to be created) through the "submitFlatbondForm"
action (also to be created, but in the next commit ;) ).
2.a. If the value is more than 120, given the challenge
constraints (and a prior form validation in the index.vue I plan),
we can be sure the user provided a weekly rent, otherwise
- the monthly one.
3. Lastly, I artificially fill either of the missing weekly/monthly
values by dividing/multiplying the monthly/weekly provided
value with 4.