(Creating first page #6) created "submitFlatbondForm" method
Method Desc:
1. This method will be called upon clicking "submit" btn.
2. It's wrapped in a conditional vuetify provides with the forms
"form.validate()" so that it executes only if form is valid.
(we will control the validation through validation rules
I'll provide in the markup)
3. It dispatches vuex's "submitFlatbondForm" with the form data.
4. It has a promise to that action call which it checks if
the submission went successfully
4.a If so, it clears the form, sets 'errorInForm" to false,
and formCreatedSuccessfully to True (this one will be
very useful)
and sets 'snackbar' to true (This will handle showing a
'toast' pop-up with message indicating form submission)
4.b. Finally with a timeout of 3 seconds (to allow for a
nice animation and a proper visual feedback to user) it
redirects to the details page (not existing page yet).