(Creating first page #7 PATCH) Added a bunch of computed properties getting data from Vuex store.
NOTICE: previous message was incidentally on the
flatbondCreatedDetailsPage.vue instead of this one.
My apologies for that. I won't revert the commit though
as it will cause the logical structure of the commits to
break and also the context of it is exactly as stated,
just the label to which page it applies is wrong.