(MAJOR:Creating first page #10) Created the whole markup for the form.
NOTICE a couple of interesting points:
- Conditional/dynamic ":Class" attributes; If the 'formCreatedSuccessful'
variable is True then various form elements will have a special class applied
to them that will make them highlighted in green (to be implemented)
- @clicks on the "__choose--box" elements; we have two btns,
depending on which we click we tell the function "chooseBoxBtnClicked"
whether it's the 'weekly' or 'monthly' arrangement btn.
That changes things in our data() variables and triggers either of
the "weekly" or "monthly" input fields to show via the "v-if" that
handles their visibility.
- the <v-icon> tags are Vuetify's built-in components that are
configured to show the Google's Material Design Icon. It'll show
the icon that the name of we specify in the tag content
(e.g. "add_circle_outline" etc.)
- ":rules" attributes in <v-text-field> components; these attributes
come baked into Vuetify's components inside the <v-form>
(e.g. v-text-field). We can specify the validation rules for the given
user's input.
NB. I have specified some of the rules inside the markup and some
(postcode rules) inside data. This is because if you specify it
inside data, you lose the "this" scope (imho that's weird) , which I need
for both the weekly and monthly rent value inputs.
- The rest should be pretty self-explanatory.