(Creating details page #2) Created "displayWeeklyOrMonthlyValue" comp. prop.
Description: this comprop evaluates
whether to show the weekly or monthly
rent amount depending on which rent
arrangement the user chose in the form.

NOTICE I made it in such a way there is
a value of 'the other' rent arr. stored in our
vuex as well, so we - PROSPECTIVELY -
we could show that as well if needs be.