(Creating details page #3) Created "flexAlignOfTwitterTitle" comp. prop.
Description: This comprop will actually be used
as a **binding** ('v-bind' attribute) to change the flexbox's parameters
on vuetify element that will hold this class once it hits the 'xs' breakpoints.
I.e. simply put - the bearer of this binding will have a
css{flex-direction: columnm align-items:center} only on a mobile viewport.
I will use it for a markup in a "Twitter card" on the page (TBC :) ).
NB I could use just CSS @media queries, but since on this page
I decided to show-off using the Vuetify's grid-system effectively,
I also decided to leverage its features to the full extent.