(Creating details page #4) Added a "snackbar" boolean variable in the "data()" of the page.
I will way of using this variable will be exactly the same as
the one used in index.vue page.
See commits: "(Creating first page #9)" for description
of the snackbar component and maybe
"(Creating first page #3)" for how the "snackbar" data is
initialised there (i.e. exactly in the same way as here)