(Creating first page #7) Created the markup for the "topbar" section of the page.
Markup description:
This section creates a stylised (actual styles are subject TBC)
section emphasizing the "flatbond created!" message.
It uses the imported "backgroundShape" component.
This component is an svg I created in Adobe XD that is intended
to nicely separate this "kinda-heading" from the background image

NOTICE at this commit stage it won't show this svg bgk shape
correctly, but it will once I add the styling in the next commits.