(Creating first page #8) Created the mockup section for displaying the Rent-amount card and the Membership-fee card info.
Section's notable points:
- I'm using Vuetify's grid system-based components
(v-layout & v-flex) therefore the markup seems
a bit more verbose than it could be (but I guess
that's the trade-off for speed of dev in this case)
- Usage of the <v-icon> tags is the same as the one
described in commit: "(MAJOR:Creating first page #10)".
- NOTICE I'm displaying the "weekly" or "monthly" value
depending on the user's choice made in the form, which I
get on this page through the "displayWeeklyOrMonthlyLabel"
and "displayWeeklyOrMonthlyValue" comp. props.