(Creating first page #10) Added a markup for a custom "twitter card" (read more below).
Markup description:
- I decided to add an extra touch to the challenge
and make a card suggesting the user to
share the fact of their flatbond created on
Twitter with a visualisation how that
tweet would look like. I also added a
dummy call to action button "share" button with it.
- NB the styling of the card will be vastly improved
once I add the styling to the whole page.
- NB.2. the usage of "twitter-____-svg" components
I imported in commit: "(Creating details page #1)"
and this is where I use them.
- NOTICE the usage of 'v-bind' attribute on "v-card-title"
component which I described in the commit:
 "(Creating details page #3)".