(FINAL: Creating details page #17) Added media queries for a correct bkg-svg-shape display on multiple viewports.
NOTICE: there still might be some missmatch of this
svg's alignment depending in the viewport size.
In real-project I would try to
seek a different solution of placing this element on
the page; the one that is less prone to such an easy
missalignment (4-5 breakpoints should usually
be enough for a visual element to align correctly
on all viewport sizes, otherwise there
probably is some more optimal solution to
place that element on the page).