Created Firebase deploy config file and deployed the web app.
This website is now available at: https://flatbond.firebaseapp.com

Notice I generated the website statically by running
"nuxt generate" and made the hosting configured
as a single-page-app.
However, in fact, I set-up the nuxt project as a universal
app (instead of the SPA mode) therefore, if needs be,
there's a room for hosting the website as a fully-fledged
SSR web app once some small configuration is done.
NB. This would become a necessity if we started
making any files in the "middleware" directory if this
project.